requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.1
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
dodgy==0.2.1
env==0.1.0
filelock==3.0.12
flake8==3.8.4
flake8-polyfill==1.0.2
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.1
pep8-naming==0.10.0
pipenv==2020.8.13
pluggy==0.13.1
prospector==1.3.1
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-decouple==3.3
python-dotenv==0.15.0
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requirements-detector==0.7
rope==0.18.0
setoptconf==0.2.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
toml==0.10.1
ujson==4.0.1
urllib3==1.25.11
virtualenv==20.0.35
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
